server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
docker
PyGithub
requests
python-dotenv
supabase
github3.py
PyJWT==2.8.0
pytest==7.4.3
pytest-mock==3.12.0